hoist four pure planning functions into codestory-agent
